tests: Move RDMACMBaseTest to be under base tests
Moving the RDMACMBaseTest to where it should as this is a base class.
